+++ stroom-core-client/src/main/java/stroom/data/client/presenter/SourcePresenter.java
@@ -510,7 +510,8 @@ public class SourcePresenter extends MyPresenterWidget<SourceView> implements Te
     private void setEditorMode(final FetchDataResult fetchDataResult) {
         final AceEditorMode mode;
 
-        if (StreamTypeNames.META.equals(fetchDataResult.getStreamTypeName())) {
+        if (fetchDataResult.getSourceLocation() != null
+                && StreamTypeNames.META.equals(fetchDataResult.getSourceLocation().getChildType())) {
             mode = AceEditorMode.PROPERTIES;
         } else {// We have no way of knowing what type the data is (could be csv, json, xml) so assume XML
             mode = AceEditorMode.XML;