+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/ReferenceDataServiceImpl.java
@@ -255,7 +255,8 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
             final LookupIdentifier lookupIdentifier = LookupIdentifier.of(
                     refDataLookupRequest.getMapName(),
                     refDataLookupRequest.getKey(),
-                    refDataLookupRequest.getEffectiveTimeEpochMs());
+                    refDataLookupRequest.getOptEffectiveTimeAsEpochMs()
+                            .orElse(Instant.now().toEpochMilli()));
 
             final List<PipelineReference> pipelineReferences = convertReferenceLoaders(
                     refDataLookupRequest.getReferenceLoaders());
@@ -288,7 +289,10 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
                 throw new NotFoundException(LogUtil.message("No value for map: {}, key: {}, time {}",
                         refDataLookupRequest.getMapName(),
                         refDataLookupRequest.getKey(),
-                        Instant.ofEpochMilli(refDataLookupRequest.getEffectiveTimeEpochMs()).toString()));
+                        refDataLookupRequest.getOptEffectiveTimeAsEpochMs()
+                        .map(Instant::ofEpochMilli)
+                        .map(Objects::toString)
+                        .orElse("null")));
             }
 
             return stringWriter.toString();